+++ application/backend/app/api/endpoints/schemas/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (C) 2025 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
