+++ application/backend/app/api/endpoints/schemas/dataset_item.py
@@ -7,8 +7,7 @@ from uuid import UUID
 from pydantic import BaseModel
 
 from app.core.models import BaseRequiredIDNameModel, Pagination
-from app.entities import DatasetItemAnnotation, DatasetItemFormat
-from app.entities.dataset_item import DatasetItemSubset
+from app.models import DatasetItemAnnotation, DatasetItemFormat, DatasetItemSubset
 
 
 class DatasetItemView(BaseRequiredIDNameModel):
@@ -58,7 +57,7 @@ class SetDatasetItemAnnotations(BaseModel):
     }
 
 
-class DatasetItemAnnotationsWithSource(BaseModel):
+class DatasetItemAnnotations(BaseModel):
     """
     Dataset item annotations with information about source
     """
