+++ application/backend/tests/integration/services/test_dataset_service.py
@@ -14,7 +14,7 @@ from sqlalchemy import func, select
 from sqlalchemy.orm import Session
 
 from app.db.schema import DatasetItemDB, DatasetItemLabelDB, PipelineDB
-from app.entities import DatasetItemAnnotation, DatasetItemSubset, LabelReference, Rectangle
+from app.models import DatasetItemAnnotation, DatasetItemSubset, LabelReference, Rectangle
 from app.schemas import PipelineView, ProjectView
 from app.services import LabelService, PipelineService, ProjectService
 from app.services.base import ResourceNotFoundError, ResourceType
