+++ application/backend/tests/unit/services/data_collect/test_data_collector.py
@@ -9,7 +9,7 @@ import numpy as np
 import pytest
 import time_machine
 
-from app.entities import DatasetItemAnnotation, DatasetItemFormat, FullImage, LabelReference
+from app.models import DatasetItemAnnotation, DatasetItemFormat, FullImage, LabelReference
 from app.schemas.pipeline import ConfidenceThresholdDataCollectionPolicy, FixedRateDataCollectionPolicy
 from app.services import DatasetService
 from app.services.data_collect.data_collector import (
