+++ application/backend/tests/unit/services/test_dataset_service.py
@@ -5,7 +5,7 @@ from uuid import uuid4
 import pytest
 
 from app.db.schema import DatasetItemDB
-from app.entities import DatasetItemAnnotation, FullImage, LabelReference, Point, Polygon, Rectangle
+from app.models import DatasetItemAnnotation, FullImage, LabelReference, Point, Polygon, Rectangle
 from app.schemas.label import LabelView
 from app.schemas.project import ProjectView, TaskType, TaskView
 from app.services import DatasetService
